+++ lib/ansible/module_utils/vmware.py
@@ -213,6 +213,24 @@ def find_host_portgroup_by_name(host, portgroup_name):
     return None
 
 
+def compile_folder_path_for_object(vobj):
+    """ make a /vm/foo/bar/baz like folder path for an object """
+
+    paths = []
+    if isinstance(vobj, vim.Folder):
+        paths.append(vobj.name)
+
+    thisobj = vobj
+    while hasattr(thisobj, 'parent'):
+        thisobj = thisobj.parent
+        if isinstance(thisobj, vim.Folder):
+            paths.append(thisobj.name)
+    paths.reverse()
+    if paths[0] == 'Datacenters':
+        paths.remove('Datacenters')
+    return '/' + '/'.join(paths)
+		
+		
 def _get_vm_prop(vm, attributes):
     """Safely get a property or return None"""
     result = vm
