+++ lib/ansible/modules/cloud/vmware/vmware_guest_find.py
@@ -80,7 +80,7 @@ import os
 # import module snippets
 from ansible.module_utils.basic import AnsibleModule
 from ansible.module_utils._text import to_native
-from ansible.module_utils.vmware import connect_to_api, gather_vm_facts, get_all_objs
+from ansible.module_utils.vmware import connect_to_api, gather_vm_facts, get_all_objs, compile_folder_path_for_object
 
 
 HAS_PYVMOMI = False
@@ -126,7 +126,7 @@ class PyVmomiHelper(object):
                 continue
             # Match by name or uuid
             if vobj.config.name == name or vobj.config.uuid == uuid:
-                folderpath = self.compile_folder_path_for_object(vobj)
+                folderpath = compile_folder_path_for_object(vobj)
                 results.append(folderpath)
 
         return results
@@ -199,24 +199,6 @@ class PyVmomiHelper(object):
             self.get_datacenter()
         self.folders = self._build_folder_tree(self.datacenter.vmFolder)
         self._build_folder_map(self.folders)
-
-    @staticmethod
-    def compile_folder_path_for_object(vobj):
-        """ make a /vm/foo/bar/baz like folder path for an object """
-
-        paths = []
-        if isinstance(vobj, vim.Folder):
-            paths.append(vobj.name)
-
-        thisobj = vobj
-        while hasattr(thisobj, 'parent'):
-            thisobj = thisobj.parent
-            if isinstance(thisobj, vim.Folder):
-                paths.append(thisobj.name)
-        paths.reverse()
-        if paths[0] == 'Datacenters':
-            paths.remove('Datacenters')
-        return '/' + '/'.join(paths)
 
     def get_datacenter(self):
         self.datacenter = get_obj(
